requirements.txt
Flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
Werkzeug==3.0.1
python-dateutil==2.8.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
